requirements.txt
torch
torchvision
monai
numpy
pydicom
scikit-learn
pillow
tk
